+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2021-10-XX
+
+- Update to Tone 14
+- Refactor `Track` to use Tone.Sequencer's `.set` because `.add` and `.remove` have been removed. Simplify setting steps.
+
 ## [0.7.0] - 2021-10-12
 
 - Remove ye olde `prop-types` in favour of `Typescript` types